+++ Openshift4/operator/artifactory-ha-operator/CHANGELOG.md
@@ -1,6 +1,19 @@
 # JFrog  Openshift Artifactory-ha Chart Changelog
 All changes to this chart will be documented in this file.
 
+## [1.1.4] - Dec 2, 2020
+* Updating Operator to latest jfrog/artifactory-ha helm chart version 4.5.4 artifactory version 7.11.5
+
+## [1.1.3] - Nov 20, 2020
+* Updating Operator to latest jfrog/artifactory-ha helm chart version 4.4.5 artifactory version 7.10.6
+
+## [1.1.2] - Oct 22, 2020
+* Updating Operator to latest jfrog/artifactory-ha helm chart version 4.4.1 artifactory version 7.10.2
+
+## [1.1.1] - Oct 6, 2020
+* Upgrade path fixed for both operator version 1.0.3 and 1.1.0
+* Updating Operator to latest jfrog/artifactory-ha helm chart version 4.2.0 artifactory version 7.9.0
+
 ## [1.1.0] - Sept 30, 2020
 * Updating Operator to latest jfrog/artifactory-ha helm chart version 4.1.0 artifactory version 7.9.0